+++ examples/client/object_connectivity_moni.c
@@ -252,7 +252,7 @@ static uint8_t prv_set_value(lwm2m_data_t * dataP,
 }
 
 static uint8_t prv_read(lwm2m_context_t *contextP,
-                        uint16_t instanceId,
+                        uint32_t instanceId,
                         int * numDataP,
                         lwm2m_data_t ** dataArrayP,
                         lwm2m_object_t * objectP)
