+++ examples/client/object_connectivity_stat.c
@@ -86,7 +86,7 @@ static uint8_t prv_set_tlv(lwm2m_data_t * dataP, conn_s_data_t * connStDataP)
 }
 
 static uint8_t prv_read(lwm2m_context_t *contextP,
-                        uint16_t instanceId,
+                        uint32_t instanceId,
                         int * numDataP,
                         lwm2m_data_t** dataArrayP,
                         lwm2m_object_t * objectP)
@@ -157,7 +157,7 @@ static void prv_resetCounter(lwm2m_object_t* objectP, bool start)
 }
 
 static uint8_t prv_exec(lwm2m_context_t *contextP,
-                        uint16_t instanceId,
+                        uint32_t instanceId,
                         uint16_t resourceId,
                         uint8_t * buffer,
                         int length,
