+++ examples/client/object_firmware.c
@@ -72,7 +72,7 @@ typedef struct
 } firmware_data_t;
 
 static uint8_t prv_firmware_read(lwm2m_context_t *contextP,
-                                 uint16_t instanceId,
+                                 uint32_t instanceId,
                                  int * numDataP,
                                  lwm2m_data_t ** dataArrayP,
                                  lwm2m_object_t * objectP)
@@ -204,7 +204,7 @@ static uint8_t prv_firmware_read(lwm2m_context_t *contextP,
 }
 
 static uint8_t prv_firmware_write(lwm2m_context_t *contextP,
-                                  uint16_t instanceId,
+                                  uint32_t instanceId,
                                   int numData,
                                   lwm2m_data_t * dataArray,
                                   lwm2m_object_t * objectP,
@@ -259,7 +259,7 @@ static uint8_t prv_firmware_write(lwm2m_context_t *contextP,
 }
 
 static uint8_t prv_firmware_execute(lwm2m_context_t *contextP,
-                                    uint16_t instanceId,
+                                    uint32_t instanceId,
                                     uint16_t resourceId,
                                     uint8_t * buffer,
                                     int length,
