+++ examples/client/object_test.c
@@ -199,7 +199,7 @@ static uint8_t prv_set_value(lwm2m_data_t * dataP,
 }
 
 static uint8_t prv_test_read(lwm2m_context_t *contextP,
-                               uint16_t instanceId,
+                               uint32_t instanceId,
                                int * numDataP,
                                lwm2m_data_t ** dataArrayP,
                                lwm2m_object_t * objectP)
@@ -260,7 +260,7 @@ static uint8_t prv_test_read(lwm2m_context_t *contextP,
 }
 
 static uint8_t prv_test_discover(lwm2m_context_t *contextP,
-                                   uint16_t instanceId,
+                                   uint32_t instanceId,
                                    int * numDataP,
                                    lwm2m_data_t ** dataArrayP,
                                    lwm2m_object_t * objectP)
@@ -345,7 +345,7 @@ static uint8_t prv_test_discover(lwm2m_context_t *contextP,
 }
 
 static uint8_t prv_test_write(lwm2m_context_t *contextP,
-                                uint16_t instanceId,
+                                uint32_t instanceId,
                                 int numData,
                                 lwm2m_data_t * dataArray,
                                 lwm2m_object_t * objectP,
@@ -419,7 +419,7 @@ static uint8_t prv_test_write(lwm2m_context_t *contextP,
 }
 
 static uint8_t prv_test_execute(lwm2m_context_t *contextP,
-                                  uint16_t instanceId,
+                                  uint32_t instanceId,
                                   uint16_t resourceId,
                                   uint8_t * buffer,
                                   int length,
